Up
Só imprime um rolo, mas associa as expsoições bem, logo poderá ser um problema de respresentação ou de nao estar a carregar os rolos como deve de ser.

+++ app/src/main/java/pt/isec/gps/grupo14/analog/AnaLog/Rolo.java
@@ -37,7 +37,7 @@ public class Rolo {
         this.idRolo = new DBHandler(context).addRolo(this);
     }
 
-    public Rolo(int id, String titulo, float iso, int formato, int maxExposicoes, String descricao,
+    public Rolo(int id, String titulo, int iso, int formato, int maxExposicoes, String descricao,
                 boolean revelado, String data, int idCamera, int nExposicoes) {
         this.idRolo = id;
         this.titulo = titulo;
